refactor: 이벤트리스너에서 도메인 로직 영역을 분리
