+++ src/js/EventListener/coveringTryCatch.js
@@ -1,0 +1,16 @@
+import ValidationError from '../ValidatorImpl/ValidationError.js';
+
+const coveringTryCatch = (tryFunction) => {
+  try {
+    tryFunction();
+  } catch (error) {
+    if (error instanceof ValidationError) {
+      error.handling();
+      return;
+    }
+
+    throw error;
+  }
+};
+
+export default coveringTryCatch;
